REALTY-DB: smoke passed for saving to DB while scraping

+++ scraping/realty-scraper-web/src/main/java/com/idealista/db/IGenericDao.java
@@ -2,7 +2,7 @@ package com.idealista.db;
 
 public interface IGenericDao<T>
 {
-    int save(T item);
+    long save(T item);
 
     T getById(long id);
 }
